Change for Product, Agent and ProductAgent entities

+++ src/main/java/com/pricecompare/repositories/AgentRepository.java
@@ -3,6 +3,6 @@ package com.pricecompare.repositories;
 import com.pricecompare.entities.Agent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentRepository extends JpaRepository<Agent, Long>
+public interface AgentRepository extends JpaRepository<Agent, Integer>
 {
 }
