+++ src/main/java/com/pricecompare/repositories/ProductRepository.java
@@ -3,6 +3,6 @@ package com.pricecompare.repositories;
 import com.pricecompare.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long>
+public interface ProductRepository extends JpaRepository<Product, Integer>
 {
 }
